+++ hotel/models/__init__.py
@@ -20,4 +20,5 @@
 ##############################################################################
 from . import hotel
 from . import hotel_folio
-from . import hotel_service
+from . import hotel_folio_room
+from . import hotel_folio_service
