+++ hotel/models/hotel_folio_room.py
@@ -5,11 +5,15 @@ from mx import DateTime
 import datetime
 from openerp.tools import config
 
-class hotel_service_line(models.Model):
+
+class HotelFolioRoom(models.Model):
+    _name = 'hotel.folio.room'
+    _description = 'Hotel Folio Room'
+    _inherits = {'sale.order.line':'order_line_id'}
     
     @api.one
     def copy(self, default=None):
-        copy = self.env['sale.order.line'].browse(id)
+        copy = self.env['sale.order.line'].browse(self.id)
         return  copy.copy(default=None)
     
     @api.multi
@@ -30,26 +34,39 @@ class hotel_service_line(models.Model):
     @api.model
     def _get_1st_packaging(self):
         return  self.env['sale.order.line']._get_1st_packaging()
+    
+    @api.model
+    def _get_checkin_date(self):
+        if 'checkin_date' in self._context:
+            return self._context['checkin_date']
+        return time.strftime('%Y-%m-%d %H:%M:%S')
+    
+    @api.model
+    def _get_checkout_date(self):
+        if 'checkin_date' in self._context:
+            return self._context['checkout_date']
+        return time.strftime('%Y-%m-%d %H:%M:%S')
+ 
+    order_line_id = fields.Many2one('sale.order.line', 'order_line_id', required=True, ondelete='cascade')
+    folio_id = fields.Many2one('hotel.folio', 'folio_id', ondelete='cascade', required=True)
+    checkin_date = fields.Datetime('Check In', required=True)
+    checkout_date = fields.Datetime('Check Out', required=True)
 
-    _name = 'hotel.service.line'
-    _description = 'hotel Service line'
-    _inherits = {'sale.order.line':'service_line_id'}
-    service_line_id = fields.Many2one('sale.order.line', 'service_line_id', required=True, ondelete='cascade')
-    folio_id = fields.Many2one('hotel.folio', 'folio_id', ondelete='cascade')
+    _defaults = {
+       'checkin_date':_get_checkin_date,
+       'checkout_date':_get_checkout_date,
+    }
     
     @api.model
     def create(self, vals, check=True):
-        if not context:
-            context = {}
-        if vals.has_key("folio_id"):
-            folio = self.env['hotel.folio'].browse([vals['folio_id']])[0]
-            vals.update({'order_id':folio.order_id.id})
-        roomline = super(models.Model, self).create(vals)
-        return roomline
-    
+        folio = self.env['hotel.folio'].browse([vals['folio_id']])[0]
+        vals.update({'order_id':folio.order_id.id})
+        return super(HotelFolioRoom, self).create(vals)
+
     @api.multi
     def uos_change(self, product_uos, product_uos_qty=0, product_id=None):
-        change = self.env['sale.order.line'].browse(self.ids)
+        line_ids = [folio.order_line_id.id for folio in self]
+        change = self.env['sale.order.line'].browse(line_ids)
         return  change.uos_change(product_uos, product_uos_qty=0, product_id=None)
     
     @api.multi
@@ -58,19 +75,19 @@ class hotel_service_line(models.Model):
             lang=False, update_tax=True, date_order=False):
         line_ids = [folio.order_line_id.id for folio in self]
         if product:
-            return self.env['sale.order.line'].browse(line_ids).product_id_change(pricelist, product, qty=0,
-            uom=False, qty_uos=0, uos=False, name='', partner_id=partner_id,
-            lang=False, update_tax=True, date_order=False)
-        
+            return self.env['sale.order.line'].browse(line_ids).product_id_change(pricelist, 
+                product, qty=0, uom=False, qty_uos=0, uos=False, name='', partner_id=partner_id,
+                lang=False, update_tax=True, date_order=False)
+    
     @api.multi    
     def product_uom_change(self, pricelist, product, qty=0,
             uom=False, qty_uos=0, uos=False, name='', partner_id=False,
             lang=False, update_tax=True, date_order=False):
         if product:
-            return  self.product_id_change(pricelist, product, qty=0,
-            uom=False, qty_uos=0, uos=False, name='', partner_id=partner_id,
-            lang=False, update_tax=True, date_order=False)
-        
+            return self.product_id_change( pricelist, product, qty=0,
+                uom=False, qty_uos=0, uos=False, name='', partner_id=partner_id,
+                lang=False, update_tax=True, date_order=False)
+
     @api.multi    
     def on_change_checkout(self, checkin_date=time.strftime('%Y-%m-%d %H:%M:%S'), checkout_date=time.strftime('%Y-%m-%d %H:%M:%S')):
         qty = 1
@@ -79,19 +96,25 @@ class hotel_service_line(models.Model):
         if checkin_date:
             diffDate = datetime.datetime(*time.strptime(checkout_date, '%Y-%m-%d %H:%M:%S')[:5]) - datetime.datetime(*time.strptime(checkin_date, '%Y-%m-%d %H:%M:%S')[:5])
             qty = diffDate.days
+            if qty == 0:
+                qty = 1
         return {'value':{'product_uom_qty':qty}}
     
     @api.multi
     def button_confirm(self):
-        button = self.env['sale.order.line'].browse(self.ids)
-        return  button.button_confirm()
+        confirm = self.env['sale.order.line'].browse(selfids)
+        return  confirm.button_confirm()
     
     @api.multi
     def button_done(self):
-        button = self.env['sale.order.line'].browse(self.ids)
-        return  button.button_done()
-    
-    @api.multi
+        res = self.env['sale.order.line'].browse(self.ids).button_done()
+        wf_service = netsvc.LocalService("workflow")
+        res = self.write({'state':'done'})
+        for line in self.browse(ids):
+            wf_service.trg_write(uid, 'sale.order', line.order_id.id, cr)
+        return res
+
+    @api.multi    
     def uos_change(self, product_uos, product_uos_qty=0, product_id=None):
         change = self.env['sale.order.line'].browse(self.ids)
         return  change.uos_change(product_uos, product_uos_qty=0, product_id=None)
@@ -99,30 +122,4 @@ class hotel_service_line(models.Model):
     @api.one
     def copy(self, default=None):
         copy = self.env['sale.order.line'].browse(self.id)
-        return copy.copy(default=None)
-
-
-class hotel_service_type(models.Model):
-    _name = "hotel.service.type"
-    _inherits = {'product.category':'ser_id'}
-    _description = "Service Type" 
-    ser_id = fields.Many2one(
-        'product.category',
-        string='Category',
-        required=True,
-        select=True,
-        ondelete='cascade')
-    _defaults = {
-        'isservicetype': lambda * a: 1,
-    }
-
-class hotel_services(models.Model):
-    
-    _name = 'hotel.services'
-    _description = 'Hotel Services and its charges'
-    _inherits = {'product.product':'service_id'}
-    service_id = fields.Many2one('product.product', string='Service', required=True, ondelete='cascade')
-    _defaults = {
-        'isservice': lambda * a: 1,
-        }
-
+        return  copy.copy(default=None)
